backend/requirements.txt
Flask==2.0.1
spacy==3.5.2
pdfminer.six==20201018
